updated Opera Test
