+++ src/test/java/suites/web/LocalChromeDriverInfoTestSuite.java
@@ -2,10 +2,11 @@ package suites.web;
 
 import org.junit.platform.suite.api.Suite;
 import org.junit.platform.suite.api.SuiteDisplayName;
+
+import ui.demo.metadata.DriverDataTest;
+
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.IncludeTags;
-
-import ui.demo.DriverDataTest;
 
 @Suite
 @SuiteDisplayName(value = "Local Chrome Driver")
