+++ src/test/java/ui/experiments/DemoOperaTest.java
@@ -5,8 +5,9 @@ import static org.junit.jupiter.api.Assertions.*;
 import java.io.File;
 import java.time.Duration;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Tags;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
@@ -16,53 +17,45 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeDriverService;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 
 import utilities.MyTestUtils;
+import utilities.WebUtils;
 
 @EnabledOnOs(OS.WINDOWS)
 public class DemoOperaTest {
 
-	WebDriver driver;
-
-	private static String url = "https://www.usa.gov/";
-
-	@BeforeEach
-	void setUp() {
-		String userDir = System.getProperty("user.dir");
-		String tempStr = userDir.substring(9);
-		int index = -1;
-		if (tempStr.contains("\\"))
-			index = tempStr.indexOf("\\");
-		else
-			index = tempStr.indexOf("/");
-		String resultStr = tempStr.substring(0, index);
-		String operaBinaryFilePath = "C:/Users/" + resultStr + "/AppData/Local/Programs/Opera/opera.exe";
-		assertTrue(new File(operaBinaryFilePath).exists(), "Failed to locate the binary.");
-
-		ChromeDriverService service = new ChromeDriverService.Builder()
-				.usingDriverExecutable(new File("src/test/resources/drivers/windows/operadriver/operadriver.exe"))
-				.build();
-		ChromeOptions options = new ChromeOptions();
-		options.setExperimentalOption("w3c", true);
-		options.setBinary(operaBinaryFilePath);
-		driver = new ChromeDriver(service, options);
-		driver.manage().window().maximize();
-		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-	}
-
-	@AfterEach
-	void tearDown() {
-		if (driver != null)
-			driver.quit();
-	}
+	private static String url = "https://jqueryui.com/";
 
 	@Test
+	@DisplayName("Demo Opera Test")
+	@Tags({ @Tag("opera") })
 	void runTest() {
+		assertTrue(MyTestUtils.isOperaOnWindowsAvailable(), "Opera browser is not available.");
+		ChromeDriverService operaService = new ChromeDriverService.Builder()
+				.usingDriverExecutable(new File("src/test/resources/drivers/win/operadriver/operadriver.exe")).build();
+		ChromeOptions operaOptions = new ChromeOptions();
+		operaOptions.setExperimentalOption("w3c", true);
+		operaOptions.setBinary(MyTestUtils.getOperaOnWindowsPath());
+		WebDriver driver = new ChromeDriver(operaService, operaOptions);
+		driver.manage().deleteAllCookies();
+		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(3));
+		driver.manage().window().maximize();
+		WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(20));
+		WebUtils utils = new WebUtils(driver);
 		driver.navigate().to(url);
-		driver.findElement(By.xpath("//li[@id='usa-nav_taxes']//a")).click();
-		MyTestUtils.pause(1);
-		assertTrue(driver.getCurrentUrl().equalsIgnoreCase("https://www.usa.gov/taxes"),
-				"User is not on the taxes page.");
+		driver.findElement(By.linkText("Tooltip")).click();
+		wait.until(ExpectedConditions.urlContains("tooltip/"));
+		driver.switchTo().frame(driver.findElement(By.className("demo-frame")));
+		By ageInputByLocator = By.id("age");
+		By tooltipContentByLocator = By.className("ui-tooltip-content");
+		utils.focusElement(driver.findElement(ageInputByLocator));
+		utils.mouse().moveToElement(driver.findElement(ageInputByLocator)).pause(Duration.ofSeconds(1)).build()
+				.perform();
+		wait.until(ExpectedConditions.visibilityOfElementLocated(tooltipContentByLocator));
+		System.out.println(driver.findElement(tooltipContentByLocator).getText());
+		driver.quit();
 	}
 
 }
