+++ src/test/java/ui/playground/PlaygroundDemoTest.java
@@ -1,6 +1,6 @@
 package ui.playground;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
